requirements.txt
numpy
PyQt5
vtk
bgem
pandas
xlrd
pygimli
psutil
attrs
